Reorganize the AL codebase for further development
