+++ source/snd_openal/snd_local.h
@@ -19,7 +19,8 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 */
 // snd_local.h -- private OpenAL sound functions
 
-
+#ifndef SND_OPENAL_LOCAL_H
+#define SND_OPENAL_LOCAL_H
 
 //#define VORBISLIB_RUNTIME // enable this define for dynamic linked vorbis libraries
 
@@ -185,226 +186,7 @@ sfx_t *S_GetBufferById( int id );
 bool S_LoadBuffer( sfx_t *sfx );
 bool S_UnloadBuffer( sfx_t *sfx );
 
-struct src_s;
 
-class Effect {
-public:
-	virtual void BindOrUpdate( src_s *src ) = 0;
-
-	virtual void InterpolateProps( const Effect *oldOne, int timeDelta ) {};
-
-	virtual unsigned GetLingeringTimeout() const {
-		return 500;
-	};
-
-	virtual bool ShouldKeepLingering( float sourceQualityHint, int64_t millisNow ) const {
-		return sourceQualityHint > 0;
-	};
-
-	virtual void UpdatePanning( src_s *src, const vec3_t listenerOrigin, const mat3_t listenerAxes ) {}
-
-	virtual ~Effect() {}
-
-	template <typename T> static const T Cast( const Effect *effect ) {
-		// We might think of optimizing it in future, that's why the method is favourable over the direct cast
-		return dynamic_cast<T>( const_cast<Effect *>( effect ) );
-	}
-
-	void AdjustGain( src_s *src ) const;
-
-	// A timestamp of last props update
-	int64_t lastUpdateAt;
-	// An distance between emitter and listener at last props update
-	float distanceAtLastUpdate;
-protected:
-	ALint type;
-	Effect( ALint type_ ): lastUpdateAt( 0 ), distanceAtLastUpdate( 0.0f ), type( type_ ) {}
-
-	class Interpolator {
-		float newWeight, oldWeight;
-	public:
-		explicit Interpolator( int timeDelta ) {
-			assert( timeDelta >= 0 );
-			float frac = timeDelta / 175.0f;
-			if( frac <= 1.0f ) {
-				newWeight = 0.5f + 0.3f * frac;
-				oldWeight = 0.5f - 0.3f * frac;
-			} else {
-				newWeight = 1.0f;
-				oldWeight = 0.0f;
-			}
-		}
-
-		float operator()( float rawNewValue, float oldValue, float mins, float maxs ) const {
-			float result = newWeight * ( rawNewValue ) + oldWeight * ( oldValue );
-			clamp( result, mins, maxs );
-			return result;
-		}
-	};
-
-	void CheckCurrentlyBoundEffect( src_s *src );
-	virtual void IntiallySetupEffect( src_s *src );
-
-	virtual float GetMasterGain( src_s *src ) const;
-
-	void AttachEffect( src_s *src );
-};
-
-class UnderwaterFlangerEffect final: public Effect {
-	void IntiallySetupEffect( src_s *src ) override;
-	float GetMasterGain( src_s *src ) const override;
-public:
-	UnderwaterFlangerEffect(): Effect( AL_EFFECT_FLANGER ) {}
-
-	float directObstruction;
-	bool hasMediumTransition;
-
-	void BindOrUpdate( src_s *src ) override;
-	void InterpolateProps( const Effect *oldOne, int timeDelta ) override;
-
-	unsigned GetLingeringTimeout() const override {
-		return 500;
-	}
-
-	bool ShouldKeepLingering( float sourceQualityHint, int64_t millisNow ) const override {
-		return sourceQualityHint > 0;
-	}
-};
-
-class ReverbEffect: public Effect {
-	float GetMasterGain( struct src_s *src ) const final;
-protected:
-	void InterpolateCommonReverbProps( const Interpolator &interpolator, const ReverbEffect *that );
-public:
-	ReverbEffect( ALint type_ ): Effect( type_ ) {}
-
-	// A regular direct obstruction (same as for the flanger effect)
-	float directObstruction;
-
-	float density;              // [0.0 ... 1.0]    default 1.0
-	float diffusion;            // [0.0 ... 1.0]    default 1.0
-	float gain;                 // [0.0 ... 1.0]    default 0.32
-	float gainHf;               // [0.0 ... 1.0]    default 0.89
-	float decayTime;            // [0.1 ... 20.0]   default 1.49
-	float reflectionsGain;      // [0.0 ... 3.16]   default 0.05
-	float reflectionsDelay;     // [0.0 ... 0.3]    default 0.007
-	float lateReverbGain;       // [0.0 ... 10.0]   default 1.26
-	float lateReverbDelay;      // [0.0 ... 0.1]    default 0.011
-
-	// An intermediate of the reverb sampling algorithm, useful for gain adjustment
-	float secondaryRaysObstruction;
-
-	virtual void CopyReverbProps( const ReverbEffect *that ) {
-		// Avoid memcpy... This is not a POD type
-		density = that->density;
-		diffusion = that->diffusion;
-		gain = that->gain;
-		gainHf = that->gainHf;
-		decayTime = that->decayTime;
-		reflectionsGain = that->reflectionsGain;
-		reflectionsDelay = that->reflectionsDelay;
-		lateReverbGain = that->lateReverbGain;
-		lateReverbDelay = that->lateReverbDelay;
-		secondaryRaysObstruction = that->secondaryRaysObstruction;
-	}
-
-	unsigned GetLingeringTimeout() const override {
-		return (unsigned)( decayTime * 1000 + 50 );
-	}
-
-	bool ShouldKeepLingering( float sourceQualityHint, int64_t millisNow ) const override {
-		if( sourceQualityHint <= 0 ) {
-			return false;
-		}
-		if( millisNow - lastUpdateAt > 200 ) {
-			return false;
-		}
-		clamp_high( sourceQualityHint, 1.0f );
-		float factor = 0.5f * sourceQualityHint;
-		factor += 0.25f * ( ( 1.0f - directObstruction ) + ( 1.0f - secondaryRaysObstruction ) );
-		assert( factor >= 0.0f && factor <= 1.0f );
-		return distanceAtLastUpdate < 192.0f + 768.0f * factor;
-	}
-};
-
-class StandardReverbEffect final: public ReverbEffect {
-	friend class ReverbEffectSampler;
-public:
-	StandardReverbEffect(): ReverbEffect( AL_EFFECT_REVERB ) {}
-
-	void BindOrUpdate( struct src_s *src ) override;
-	void InterpolateProps( const Effect *oldOne, int timeDelta ) override;
-};
-
-class EaxReverbEffect final: public ReverbEffect {
-	friend class ReverbEffectSampler;
-
-	const float *MakeFakeSourceOrigin( struct src_s *src,
-									   const vec3_t listenerOrigin,
-									   const vec3_t *reflectionDirs,
-									   unsigned numReflectionDirs );
-
-	void UpdateDelegatedSpatialization( struct src_s *src,
-										const vec3_t listenerOrigin,
-										const vec3_t *reflectionDirs,
-										unsigned numReflectionDirs );
-
-	vec3_t tmpSourceOrigin { 0, 0, 0 };
-public:
-	EaxReverbEffect(): ReverbEffect( AL_EFFECT_EAXREVERB ) {}
-
-	float echoTime;   // [0.075 ... 0.25]  default 0.25
-	float echoDepth;  // [0.0   ... 1.0]   default 0.0
-
-	void BindOrUpdate( struct src_s *src ) override;
-	void InterpolateProps( const Effect *oldOne, int timeDelta ) override;
-
-	void CopyReverbProps( const ReverbEffect *that ) override;
-
-	void UpdatePanning( src_s *src, const vec3_t listenerOrigin, const mat3_t listenerAxes ) override;
-};
-
-class ChorusEffect final: public Effect {
-public:
-	ChorusEffect(): Effect( AL_EFFECT_CHORUS ) {}
-
-	enum Waveform {
-		SIN = 0,
-		TRIANGLE = 1
-	};
-
-	Waveform waveform;
-	int phase;        // [-180 ... 180]     default 90
-	float rate;       // [0.0  ... 10.0]    default 1.1
-	float depth;      // [0.0  ... 1.0]     default 0.1
-	float feedback;   // [-1.0 ... 1.0]     default 0.25
-	float delay;      // [0.0  ... 0.016]   default 0.016
-
-	void BindOrUpdate( struct src_s *src ) override;
-};
-
-class DistortionEffect final: public Effect {
-public:
-	DistortionEffect(): Effect( AL_EFFECT_DISTORTION ) {}
-
-	float edge;   // [0.0  ... 1.0]  default 0.2
-	float gain;   // [0.01 ... 1.0]  default 0.05
-
-	void BindOrUpdate( struct src_s *src ) override;
-};
-
-class EchoEffect final: public Effect {
-public:
-	EchoEffect(): Effect( AL_EFFECT_ECHO ) {}
-
-	float delay;      // [0.0  ... 0.207]  default 0.1
-	float lrDelay;    // [0.0  ... 0.404]  default 0.1
-	float damping;    // [0.0  ... 0.99]   default 0.5
-	float feedback;   // [0.0  ... 1.0]    default 0.5
-	float spread;     // [-1.0 ... 1.0]    default -1.0
-
-	void BindOrUpdate( struct src_s *src ) override;
-};
 
 typedef struct {
 	float quality;
@@ -418,6 +200,8 @@ struct PanningUpdateState {
 	vec3_t reflectionPoints[MAX_POINTS];
 	unsigned numReflectionPoints;
 };
+
+class Effect;
 
 typedef struct envUpdateState_s {
 	sfx_t *parent;
@@ -602,4 +386,6 @@ void SF_PositionedRawSamples( int entnum, float fvol, float attenuation,
 
 #ifdef __cplusplus
 }
+#endif
+
 #endif
